requirements.txt
accelerate
pandas
pydub
pyannote-audio
pytest
python-dotenv
torch
tqdm
transformers
